backend/api_gateway/requirements.txt
Django==5.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
celery==5.6.3
redis==7.4.0
djangorestframework-simplejwt==5.5.1
python-dotenv==1.2.2
django-storages==1.14.6
boto3==1.42.85
psycopg2-binary==2.9.11
markdown==3.10.2
django-filter==25.2
django-allauth==65.16.0
dj-rest-auth==7.2.0